Refactor auth into plug
